made changes

+++ ft_bzero.c
@@ -3,16 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   ft_bzero.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mknoll <mknoll@student.42.fr>              +#+  +:+       +#+        */
+/*   By: moritzknoll <moritzknoll@student.42.fr>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/15 12:23:28 by mknoll            #+#    #+#             */
-/*   Updated: 2024/10/22 10:52:43 by mknoll           ###   ########.fr       */
+/*   Updated: 2024/10/22 14:42:14 by moritzknoll      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	bzero(void *s, int n)
+#include <stdio.h>
+
+void	bzero(void *s, size_t n)
 {
-	int				i;
+	unsigned int	i;
 	unsigned char	*byte_ptr;
 
 	i = 0;
